requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
dj-static==0.0.6
Django==5.2.4
django-cors-headers==4.7.0
django-ninja==1.4.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
static3==0.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
